+++ src/CP/Utilities/CoreUtilities.php
@@ -70,6 +70,7 @@ class CoreUtilities
             ->docsUrl(Statamic::docsUrl('licensing'))
             ->routes(function ($router) {
                 $router->get('refresh', [LicensingController::class, 'refresh'])->name('refresh');
+                $router->post('mint', [LicensingController::class, 'mint'])->name('mint');
             });
 
         if (config('statamic.git.enabled') && Statamic::pro()) {